ingestion_app/requirements.txt
Flask==3.0.2
snowflake-connector-python==3.12.1
python-dotenv==1.0.1